New validate command
