+++ .gitignore
@@ -1,3 +1,4 @@
+__pycache__/
 servers.json
 nginx.conf
 nginx.conf.bak
